+++ src/sections/dataset/dataset-files/files-table/file-actions/file-actions-cell/file-action-buttons/FileActionButtons.tsx
@@ -31,9 +31,10 @@ export function FileActionButtons({
         userHasDownloadPermission={file.permissions.canDownloadFile}
         metadata={file.metadata}
         isDeaccessioned={file.datasetPublishingStatus === DatasetPublishingStatus.DEACCESSIONED}
+        isDraft={file.datasetPublishingStatus === DatasetPublishingStatus.DRAFT}
+        canEdit={file.permissions.canEditOwnerDataset}
         ingestInProgress={file.ingest.isInProgress}
         asIcon
-        isDraft={file.datasetPublishingStatus === DatasetPublishingStatus.DRAFT}
       />
       <FileOptionsMenu
         file={file}