+++ src/sections/edit-dataset-terms/EditDatasetTermsFactory.tsx
@@ -6,9 +6,11 @@ import { DatasetJSDataverseRepository } from '@/dataset/infrastructure/repositor
 import { DatasetProvider } from '../dataset/DatasetProvider'
 import { ReactElement } from 'react'
 import { DatasetNonNumericVersion } from '@/dataset/domain/models/Dataset'
+import { GuestbookJSDataverseRepository } from '@/guestbooks/infrastructure/repositories/GuestbookJSDataverseRepository'
 
 const licenseRepository = new LicenseJSDataverseRepository()
 const datasetRepository = new DatasetJSDataverseRepository()
+const guestbookRepository = new GuestbookJSDataverseRepository()
 
 export class EditDatasetTermsFactory {
   static create(): ReactElement {
@@ -31,6 +33,7 @@ function EditDatasetTermsWithSearchParams() {
         defaultActiveTabKey={defaultActiveTabKey}
         licenseRepository={licenseRepository}
         datasetRepository={datasetRepository}
+        guestbookRepository={guestbookRepository}
       />
     </DatasetProvider>
   )