+++ src/sections/not-implemented/NotImplementedModal.tsx
@@ -17,7 +17,7 @@ export function NotImplementedModal({ show, handleClose }: NotImplementedModalPr
         <Modal.Title>{modalTitle}</Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        <p>This feature is not implemented yet in SPA.</p>
+        <p>This feature is not implemented yet in the Modern version.</p>
         <p>
           If you want to use this feature you can go to the original{' '}
           <a href={baseUrl}>Dataverse page</a>.